XD-120 - Move top level Main classes to sub-package 'server'
