+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/server/ContainerMain.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.xd;
+package org.springframework.xd.dirt.server;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
