+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/server/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Contains the Main classes for Admin and Container.
+ *
+ */
+package org.springframework.xd.dirt.server;
